+++ commhandler/src/sgr_commhandler/driver/rest/restapi_interface_async.py
@@ -1,9 +1,8 @@
 import json
 import logging
+import re
 import ssl
-from io import UnsupportedOperation
-from typing import Any, Callable, Optional
-from urllib.parse import urlencode
+from typing import Any, Optional
 
 import aiohttp
 import certifi
@@ -43,9 +42,14 @@ from sgr_commhandler.api.dynamic_parameter import (
     build_dynamic_parameters,
     build_dynamic_parameter_substitutions
 )
+from sgr_commhandler.driver.rest.request import (
+    RestRequest,
+    RestResponse,
+    build_rest_request
+)
 from sgr_commhandler.driver.rest.authentication import setup_authentication
 from sgr_commhandler.validators import build_validator
-from sgr_commhandler.utils import jmespath_mapping
+from sgr_commhandler.utils import jmespath_mapping, template
 
 logger = logging.getLogger(__name__)
 
@@ -61,36 +65,6 @@ def build_rest_data_point(
         data_type = data_point.data_point.data_type
     validator = build_validator(data_type)
     return DataPoint(protocol, validator)
-
-
-def substitute_parameters(template: str, substitutions: dict[str, str]) -> str:
-    conv_template = template
-    for name, value in substitutions.items():
-        conv_template = conv_template.replace(f'[[{name}]]', value)
-    return conv_template
-
-class RestResponse:
-    def __init__(self, headers: CIMultiDict[str] = CIMultiDict(), body: Optional[str] = None):
-        self.headers = headers
-        self.body = body
-
-
-class RestRequest:
-    def __init__(
-        self,
-        method: str,
-        url: str,
-        headers: CIMultiDict[str] = CIMultiDict(),
-        query_parameters: CIMultiDict[str] = CIMultiDict(),
-        form_parameters: CIMultiDict[str] = CIMultiDict(),
-        body: Optional[str] = None,
-    ):
-        self.method = method
-        self.url = url
-        self.headers = headers
-        self.query_parameters = query_parameters
-        self.form_parameters = form_parameters
-        self.body = body
 
 
 class RestDataPoint(DataPointProtocol[RestApiDataPointSpec]):
@@ -221,7 +195,11 @@ class RestDataPoint(DataPointProtocol[RestApiDataPointSpec]):
         if not self._read_call and not self._write_call:
             raise Exception('No REST service call configured')
         
-        self._dynamic_parameters = build_dynamic_parameters(self._dp_spec.data_point.parameter_list)
+        self._dynamic_parameters = build_dynamic_parameters(
+            self._dp_spec.data_point.parameter_list
+            if self._dp_spec.data_point
+            else None
+        )
 
         self._fp_name = ''
         if (
@@ -245,35 +223,43 @@ class RestDataPoint(DataPointProtocol[RestApiDataPointSpec]):
     def get_specification(self) -> RestApiDataPointSpec:
         return self._dp_spec
 
-    async def get_val(self, parameters: Optional[dict[str, str]] = None, skip_cache: bool = False):
+    async def get_val(self, parameters: Optional[dict[str, str]] = None, skip_cache: bool = False) -> Any:
         if not self._read_call:
             raise Exception('No read call')
 
         substitutions = build_dynamic_parameter_substitutions(self._dynamic_parameters, parameters)
 
-        request = self._build_request(self._read_call, substitutions)
+        request = build_rest_request(self._read_call, str(self._interface.base_url), substitutions)
 
         response = await self._interface.execute_request(request, skip_cache)
         if not response.body:
             return None
         if (
             self._read_call.response_query
-            and self._read_call.response_query.query_type
-            == ResponseQueryType.JMESPATH_EXPRESSION
+            and self._read_call.response_query.query_type == ResponseQueryType.JMESPATH_EXPRESSION
         ):
-            query_expression = substitute_parameters(
+            # JMESPath expression
+            query_expression = template.substitute(
                 self._read_call.response_query.query if self._read_call.response_query.query else '',
                 substitutions
             )
             return jmespath.search(query_expression, json.loads(response.body))
         elif (
             self._read_call.response_query
-            and self._read_call.response_query.query_type
-            == ResponseQueryType.JMESPATH_MAPPING
+            and self._read_call.response_query.query_type == ResponseQueryType.JMESPATH_MAPPING
         ):
             # JMESPath mappings
             mappings = self._read_call.response_query.jmes_path_mappings.mapping if self._read_call.response_query.jmes_path_mappings else []
             return jmespath_mapping.map_json_response(response.body, mappings)
+        elif (
+            self._read_call.response_query
+            and self._read_call.response_query.query_type == ResponseQueryType.REGULAR_EXPRESSION
+        ):
+            # regex
+            query_expression = self._read_call.response_query.query if self._read_call.response_query.query else ''
+            query_match = re.match(query_expression, response.body)
+            if query_match is not None:
+                return query_match.group()
 
         # plain response
         ret_value = response.body
@@ -281,7 +267,7 @@ class RestDataPoint(DataPointProtocol[RestApiDataPointSpec]):
         # apply value mappings
         if self._read_call.value_mapping:
             mappings = self._read_call.value_mapping.mapping
-            for m in mappings.items():
+            for m in mappings:
                 if m.device_value == ret_value:
                     ret_value = m.generic_value
                     break
@@ -293,7 +279,7 @@ class RestDataPoint(DataPointProtocol[RestApiDataPointSpec]):
             and self._dp_spec.data_point.unit_conversion_multiplicator != 1.0
         ):
             ret_value = (
-                float(ret_value)
+                float(str(ret_value))
                 * self._dp_spec.data_point.unit_conversion_multiplicator
             )
 
@@ -315,7 +301,7 @@ class RestDataPoint(DataPointProtocol[RestApiDataPointSpec]):
         value = str(value)
         if self._read_call.value_mapping:
             mappings = self._read_call.value_mapping.mapping
-            for m in mappings.items():
+            for m in mappings:
                 if m.generic_value == value:
                     value = m.device_value
                     break
@@ -323,7 +309,7 @@ class RestDataPoint(DataPointProtocol[RestApiDataPointSpec]):
         substitutions = {
             'value': value
         }
-        request = self._build_request(self._write_call, substitutions=substitutions)
+        request = build_rest_request(self._write_call, str(self._interface.base_url), substitutions=substitutions)
         # TODO use response body
         await self._interface.execute_request(request, skip_cache=True)
 
@@ -345,68 +331,6 @@ class RestDataPoint(DataPointProtocol[RestApiDataPointSpec]):
 
     def dynamic_parameters(self) -> list[DynamicParameter]:
         return self._dynamic_parameters
-
-    def subscribe(self, fn: Callable[[Any], None]):
-        raise UnsupportedOperation(
-            'subscribe is no available for rest data point'
-        )
-
-    def unsubscribe(self):
-        raise UnsupportedOperation(
-            'unsubscribe is no available for rest data point'
-        )
-
-    def _build_request(self, call_spec: RestApiServiceCall, substitutions: dict[str, str]) -> RestRequest:
-
-        # copy from DP spec.
-        method = call_spec.request_method if call_spec.request_method else HttpMethod.GET
-        req_path = str(call_spec.request_path) if call_spec.request_path else ''
-        headers = call_spec.request_header if call_spec.request_header else HeaderList()
-        query = call_spec.request_query if call_spec.request_query else ParameterList()
-        form = call_spec.request_form if call_spec.request_form else ParameterList()
-        body=str(call_spec.request_body) if call_spec.request_body else None
-
-        # All headers into dictionary, with substitution
-        request_headers: CIMultiDict[str] = CIMultiDict()
-        for header_entry in headers.header:
-            if header_entry.header_name:
-                request_headers.add(header_entry.header_name, substitute_parameters(header_entry.value, substitutions))
-
-        # All query parameters into dictionary, with substitution
-        query_parameters: CIMultiDict[str] = CIMultiDict()
-        for param_entry in query.parameter:
-            if param_entry.name:
-                query_parameters.add(param_entry.name, substitute_parameters(param_entry.value, substitutions))
-
-        # All form parameters into dictionary, with substitution
-        form_parameters: CIMultiDict[str] = CIMultiDict()
-        for param_entry in form.parameter:
-            if param_entry.name:
-                form_parameters.add(param_entry.name, substitute_parameters(param_entry.value, substitutions))
-
-        # request path, with substitution
-        request_path = substitute_parameters(req_path, substitutions)
-
-        # request body, with substitution
-        request_body: Optional[str] = substitute_parameters(body, substitutions) if body is not None else None
-
-        # override body when form parameters are set
-        if len(form_parameters) > 0:
-            request_body = urlencode(form_parameters)
-            request_headers['Content-Type'] = (
-                'application/x-www-form-urlencoded'
-            )
-
-        base_url = f'{self._interface.base_url}{request_path}'
-
-        return RestRequest(
-            method.value,
-            base_url,
-            request_headers,
-            query_parameters,
-            form_parameters,
-            request_body
-        )
 
 
 class RestFunctionalProfile(FunctionalProfile):
@@ -459,7 +383,7 @@ class SGrRestInterface(SGrBaseInterface):
     def __init__(
         self, frame: DeviceFrame
     ):
-        self._initialize_device(frame)
+        super().__init__(frame)
         self._session = None
         self._cache = TTLCache(maxsize=100, ttl=5)
 
@@ -529,7 +453,7 @@ class SGrRestInterface(SGrBaseInterface):
                 data=request.body
             ) as req:
                 req.raise_for_status()  # Raises an HTTPError if the HTTP request returned an unsuccessful status code
-                logger.info(f'execute_request status: {req.status}')
+                logger.debug(f'execute_request status: {req.status}')
                 res_body = await req.text()
 
                 res_headers = CIMultiDict()